requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6
Babel>=1.3
python-novaclient>=2.26.0
networkx>=1.10
oslo.log>=1.12.0 # Apache-2.0
oslo.policy>=0.3.0
pecan>=0.8.0
PasteDeploy>=1.5.0
Werkzeug>=0.7
keystonemiddleware>=2.3.0
stevedore>=1.5.0 # Apache-2.0
exrex>=0.9.4
